refactored api client
